requirements.txt
Flask==3.0.3
# langchain_core==0.3.6
# langchain_groq==0.2.0
openai==1.48.0
pandas==2.2.3
psycopg2_binary==2.9.9
PyYAML==6.0.2